+++ src/lib/server/data/time.ts
@@ -1,6 +1,6 @@
-import type { OptActionState, TimeEntryRecord, UserInfo, UserSchedule } from '$lib/schema';
+import type { OptActionState, TimeEntryRecord, UserInfo, UserSchedule } from '$lib/types/schema';
 import type { Session } from 'lucia';
-import { db } from "$lib/server/database/db-controller";
+import { db } from '$lib/server/database/db-controller';
 import { getSchedule } from './schedule';
 import { getWeekRange } from '$lib/utility';
 
@@ -15,42 +15,25 @@ export const postTime = async (
     timeAction: OptActionState;
   }
 ) => {
-  if (data.category == 'clock') {
-    if (data.timeAction == 'start') {
-      return await db.clockIn({
-        user_id: data.user_id,
-        category: data.category,
-        date_at: data.date_at,
-        sched_id: data.sched_id,
-        user_ip: data.user_ip,
-        user_agent: data.user_agent,
-        start_at: data.timestamp,
-        remarks: data.remarks
-      });
-    } else {
-      return await db.clockOut({
-        id: data.id,
-        end_at: data.timestamp
-      });
-    }
+  if (data.timeAction == 'start') {
+    return await db.startTime({
+      user_id: data.user_id,
+      category: data.category,
+      date_at: data.date_at,
+      sched_id: data.sched_id,
+      start_at: data.timestamp,
+      user_ip: data.user_ip,
+      user_agent: data.user_agent,
+      remarks: data.remarks
+    });
   } else {
-    if (data.timeAction == 'start') {
-      return await db.startTime({
-        user_id: data.user_id,
-        category: data.category,
-        date_at: data.date_at,
-        sched_id: data.sched_id,
-        start_at: data.timestamp,
-        remarks: data.remarks
-      });
-    } else {
-      return await db.endTime({
-        id: data.id,
-        end_at: data.timestamp,
-        user_ip: data.user_ip,
-        user_agent: data.user_agent
-      });
-    }
+    return await db.endTime({
+      id: data.id,
+      end_at: data.timestamp,
+      user_ip: data.user_ip,
+      user_agent: data.user_agent,
+      remarks: data.remarks
+    });
   }
 };
