+++ src/lib/server/database/sql-queries.ts
@@ -1,4 +1,4 @@
-import type { TimeEntryRecord } from '$lib/schema';
+import type { TimeEntryRecord } from '$lib/types/schema';
 
 export const QUERY = {
   USER: (args: { user_id: number }) => {
@@ -7,12 +7,25 @@ export const QUERY = {
       args
     };
   },
-  USERS_LIST: (args: { region?: string, lead_id?: number, active?: number, last_id?: number, limit?: number, }) => {
+
+  USERS_LIST: (args: {
+    region?: string;
+    lead_id?: number;
+    active?: number;
+    last_id?: number;
+    limit?: number;
+  }) => {
     const where = [];
-    const values: { region?: string, limit?: number, lead_id?: number, active?: number, last_id?: number } = {
+    const values: {
+      region?: string;
+      limit?: number;
+      lead_id?: number;
+      active?: number;
+      last_id?: number;
+    } = {
       last_id: args.last_id || 0,
       limit: args.limit || 10
-    }
+    };
     if (args.region) {
       where.push('region = $region');
       values.region = args.region;
@@ -29,16 +42,16 @@ export const QUERY = {
     }
 
     return {
-      sql: `SELECT * FROM users_list WHERE id > $last_id ${(where.length ? `AND ${where.join(' AND ')}` : '') } LIMIT $limit`,
+      sql: `SELECT * FROM users_list WHERE id > $last_id ${where.length ? `AND ${where.join(' AND ')}` : ''} LIMIT $limit`,
       args: values
-    }
+    };
   },
   USER_SCHEDULES: (args: { user_id: number; limit?: number }) => {
     return {
       sql: `SELECT 
-				date(current_timestamp, CONCAT(local_offset, ' hours')) as date_at, * 
-				FROM schedules WHERE effective_date <= date_at AND user_id = $user_id
-				ORDER BY effective_date DESC LIMIT $limit`,
+	      date(current_timestamp, CONCAT(local_offset, ' hours')) as date_at, * 
+	      FROM schedules WHERE effective_date <= date_at AND user_id = $user_id
+	      ORDER BY effective_date DESC LIMIT $limit`,
       args: {
         user_id: args.user_id,
         limit: !args.limit ? 10 : args.limit
@@ -57,9 +70,9 @@ export const QUERY = {
   LAST_ENTRY: (args: { user_id: number }) => {
     return {
       sql: `WITH recent AS 
-							(SELECT MAX(date_at) as date_at, user_id FROM time_entries WHERE user_id = $user_id AND category = 'clock')
-						SELECT e.* FROM time_entries e, recent r 
-						WHERE e.date_at = r.date_at and e.user_id = r.user_id`,
+	      (SELECT MAX(date_at) as date_at, user_id FROM time_entries WHERE user_id = $user_id AND category = 'clock')
+	      SELECT e.* FROM time_entries e, recent r 
+	    WHERE e.date_at = r.date_at and e.user_id = r.user_id`,
       args
     };
   },
@@ -78,14 +91,36 @@ export const QUERY = {
 };
 
 export const WRITE = {
-  CLOCKIN: (args: Omit<TimeEntryRecord, 'id' | 'end_at' | 'elapse_sec' | 'total_sec'>) => {
+  STARTTIME: (args: Omit<TimeEntryRecord, 'id' | 'end_at' | 'elapse_sec' | 'total_sec'>) => {
     return {
-      sql: `INSERT INTO time_entries (user_id, sched_id, category, date_at, start_at, user_ip, user_agent, remarks)
-				VALUES ($user_id, $sched_id, $category, $date_at, $start_at, $user_ip, $user_agent, $remarks) RETURNING *`,
+      sql: `INSERT INTO time_entries 
+	      (user_id, sched_id, category, date_at, start_at, user_ip, user_agent, remarks)
+	      VALUES ($user_id, $sched_id, $category, $date_at, $start_at, $user_ip, $user_agent, $remarks) RETURNING *`,
       args
     };
   },
-  CLOCKOUT: (args: Pick<TimeEntryRecord, 'id' | 'end_at'>) => {
+  ENDTIME: (
+    args: Pick<TimeEntryRecord, 'id' | 'end_at' | 'user_ip' | 'user_agent' | 'remarks'>
+  ) => {
+    return {
+      sql: `UPDATE time_entries SET 
+	      end_at = $end_at,
+	      user_ip = CONCAT(user_ip, ' | ', $user_ip),
+	      user_agent = CONCAT(user_agent, ' | ', $user_agent),
+	      remarks = NULLIF(CONCAT(remarks || ' | ', $remarks), '')
+	      WHERE id = $id RETURNING *`,
+      args
+    };
+  },
+  CLOCKIN: (args: Omit<TimeEntryRecord, 'id' | 'end_at' | 'elapse_sec' | 'total_sec'>) => {
+    return {
+      sql: `INSERT INTO time_entries 
+	      (user_id, sched_id, category, date_at, start_at, user_ip, user_agent, remarks)
+	      VALUES ($user_id, $sched_id, $category, $date_at, $start_at, $user_ip, $user_agent, $remarks) RETURNING *`,
+      args
+    };
+  },
+  CLOCKOUT: (args: Pick<TimeEntryRecord, 'id' | 'end_at' | 'user_id' | 'user_agent'>) => {
     return {
       sql: `UPDATE time_entries SET end_at = $end_at WHERE id = $id RETURNING *`,
       args
@@ -96,7 +131,7 @@ export const WRITE = {
   ) => {
     return {
       sql: `INSERT INTO time_entries (user_id, sched_id, category, date_at, start_at, remarks)
-				VALUES ($user_id, $sched_id, $category, $date_at, $start_at, $remarks) RETURNING *`,
+	      VALUES ($user_id, $sched_id, $category, $date_at, $start_at, $remarks) RETURNING *`,
       args
     };
   },
