+++ src/routes/(private)/+layout.server.ts
@@ -1,4 +1,3 @@
-import { redirect } from '@sveltejs/kit';
 import type { LayoutServerLoad } from './$types';
 
 export const load: LayoutServerLoad = async ({ locals }) => {
