+++ src/routes/(public)/login/+page.server.ts
@@ -3,7 +3,6 @@ import { lucia } from '$lib/server/lucia/auth';
 import { fail, redirect } from '@sveltejs/kit';
 import { validateSignIn } from '$lib/validation';
 import { validateUser } from '$lib/server/data/user';
-import { timeProcess } from '$lib/utility';
 
 export const load: PageServerLoad = async ({ locals }) => {
   if (locals?.session) {
@@ -22,9 +21,7 @@ export const actions = {
       return fail(400);
     }
 
-    timeProcess()
     const user = await validateUser(inputValid.data);
-    timeProcess('User Validation')
     if (!user) {
       return fail(400);
     }
@@ -39,7 +36,6 @@ export const actions = {
       path: '.',
       ...sessionCookie.attributes
     });
-
 
     redirect(302, `/user/${user.id}/timesheets`);
   }
